+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 ## Unreleased
 
 ### 4.3.0
+- Update about-us page to include staffing change
 - Bug fix: Use `ogImage` rather than `leadingImage` as og image on topic landing page
-
 - Remove legacy code
-- Update Webpack `resolve.modules` to retrieve modules from the `node_modules` of this repo
+
+#### Update webpack.config.js
+- Update `process.env` format for `webpack.DefinePlugin`
+- Update `resolve.modules` to retrieve modules from the `node_modules` of this repo firstly
 
 #### Adapt new packages
 - Add new dependencies `@twreporter/core` and `@twreporter/index-page`
@@ -22,6 +25,7 @@
 - Bump `@twreporter/redux` version to 5.0.1
 
 ## Release
+
 ### 4.2.1 
 #### Dependency Update 
 - @twreporter/react-components@^6.1.4